delivery-robot/backend/requirements.txt
Django>=5.2,<6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
channels==4.2.0
channels-redis==4.2.1
psycopg[binary]>=3.2.10
redis==5.2.1
python-dotenv==1.0.1
uvicorn==0.34.0
